+++ src/machine.rs
@@ -1,6 +1,10 @@
 //! The crypto specific Olm objects.
 
-use std::{collections::BTreeMap, ops::Deref, time::Duration};
+use std::{
+    collections::{BTreeMap, HashSet},
+    ops::Deref,
+    time::Duration,
+};
 
 use futures_util::{pin_mut, StreamExt};
 use js_sys::{Array, Function, JsString, Map, Promise, Set};
@@ -32,7 +36,7 @@ use crate::{
     store,
     store::{RoomKeyInfo, StoreHandle},
     sync_events,
-    types::{self, RoomKeyImportResult, SignatureVerification},
+    types::{self, RoomKeyImportResult, RoomSettings, SignatureVerification},
     verification, vodozemac,
 };
 
@@ -1267,6 +1271,26 @@ impl OlmMachine {
         });
     }
 
+    /// Register a callback which will be called whenever there is an update to
+    /// a device.
+    ///
+    /// `callback` should be a function that takes a single argument (an array
+    /// of user IDs as strings) and returns a Promise.
+    #[wasm_bindgen(js_name = "registerDevicesUpdatedCallback")]
+    pub async fn register_devices_updated_callback(&self, callback: Function) {
+        let stream = self.inner.store().identities_stream_raw();
+
+        // fire up a promise chain which will call `callback` on each result from the
+        // stream
+        spawn_local(async move {
+            // take a reference to `callback` (which we then pass into the closure), to stop
+            // the callback being moved into the closure (which would mean we could only
+            // call the closure once)
+            let callback_ref = &callback;
+            stream.for_each(move |item| send_device_updates_to_callback(callback_ref, item)).await;
+        });
+    }
+
     /// Register a callback which will be called whenever a secret
     /// (`m.secret.send`) is received.
     ///
@@ -1379,6 +1403,42 @@ impl OlmMachine {
         })
     }
 
+    /// Get the stored room settings, such as the encryption algorithm or
+    /// whether to encrypt only for trusted devices.
+    ///
+    /// These settings can be modified via {@link #setRoomSettings}.
+    ///
+    /// # Returns
+    ///
+    /// `Promise<RoomSettings|undefined>`
+    #[wasm_bindgen(js_name = "getRoomSettings")]
+    pub async fn get_room_settings(
+        &self,
+        room_id: &identifiers::RoomId,
+    ) -> Result<JsValue, JsError> {
+        let result = self.inner.room_settings(&room_id.inner).await?;
+        Ok(result.map(|settings| RoomSettings::from(settings)).into())
+    }
+
+    /// Store encryption settings for the given room.
+    ///
+    /// This method checks if the new settings are "safe" -- ie, that they do
+    /// not represent a downgrade in encryption security from any previous
+    /// settings. Attempts to downgrade security will result in an error.
+    ///
+    /// If the settings are valid, they will be persisted to the crypto store.
+    /// These settings are not used directly by this library, but the saved
+    /// settings can be retrieved via {@link #getRoomSettings}.
+    #[wasm_bindgen(js_name = "setRoomSettings")]
+    pub async fn set_room_settings(
+        &self,
+        room_id: &identifiers::RoomId,
+        room_settings: &RoomSettings,
+    ) -> Result<(), JsError> {
+        self.inner.set_room_settings(&room_id.inner, &room_settings.into()).await?;
+        Ok(())
+    }
+
     /// Shut down the `OlmMachine`.
     ///
     /// The `OlmMachine` cannot be used after this method has been called.
@@ -1439,6 +1499,34 @@ async fn send_user_identities_to_callback(
             Err(e) => {
                 warn!("Error calling user-identity-updated callback: {:?}", e);
             }
+        }
+    }
+}
+
+// helper for register_device_updated_callback: passes the user IDs into
+// the javascript function
+async fn send_device_updates_to_callback(
+    callback: &Function,
+    (_, device_updates): (IdentityChanges, DeviceChanges),
+) {
+    // get the user IDs of all the devices that have changed
+    let updated_chain = device_updates
+        .new
+        .into_iter()
+        .chain(device_updates.changed.into_iter())
+        .chain(device_updates.deleted.into_iter());
+    // put them in a set to make them unique
+    let updated_users: HashSet<String> =
+        HashSet::from_iter(updated_chain.map(|device| device.user_id().to_string()));
+    let updated_users_vec = Vec::from_iter(updated_users.iter());
+    match promise_result_to_future(
+        callback.call1(&JsValue::NULL, &serde_wasm_bindgen::to_value(&updated_users_vec).unwrap()),
+    )
+    .await
+    {
+        Ok(_) => (),
+        Err(e) => {
+            warn!("Error calling device-updated callback: {:?}", e);
         }
     }
 }